feat: move buttons to panel, add toggle
