+++ main.css
@@ -1,3 +1,4 @@
+
 /* all main styles here */
 
 #header-container {
@@ -17,27 +18,16 @@ a {
     font-family: 'Avenir', sans-serif;
 }
 
-.container {
-    margin: 0 auto;
-    min-height: 100vh;
-    height: auto;
-}
-
 
 #container {
     margin: 0 auto;
-    transform-origin: top center; 
+    transform-origin: top left; 
     transform: scale(1);
     transition: transform 0.3s ease;
     min-height: 100vh;
     height: auto;
 }
 
-
-.body {
-    min-height: 100vh;
-    height: auto;
-}
 
 .modal {
     display: none;
@@ -60,6 +50,20 @@ a {
     width: 80%;
 }
 
+.control {
+    margin-bottom: 1em;
+}
+
+.control button{
+    margin-right: 1em;
+}
+
+#editor {
+    display: none;
+    height: 95dvh;
+    width: 98dvw;
+}
+
 .close {
     color: #aaa;
     float: right;
@@ -74,31 +78,69 @@ a {
     cursor: pointer;
 }
 
-#zoom-in-button, #zoom-out-button, #reset-button {
-    background-color: transparent;
-    border: none;
-    padding: 0;
-    cursor: pointer;
-    position: fixed; 
-    right: 48px;
-}
-
-:root{
-    --thumbnail-width: 10%;
-}
-
+/* styles for interaction panel */
 #interaction-panel{
     position: fixed;
+    padding: 5px;
     background-color: rgba(255, 255, 255, 0.6);
     top: 20px;
-    right: 40px;
-    width: 30px;
-    height: 120px;
+    right: 33px;
+    width: auto;
+    height: auto;
     border: 1px solid #cccccc6d;
     border-radius: 5px;
     box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
     z-index: 30;
 }
+
+.panelButtons {
+    background-color: transparent;
+    border: none;
+    padding-top: 5px;
+    padding-bottom: 5px;
+    cursor: pointer;
+    position: relative;
+    /* display: inline-block; */
+    display: flex;
+    align-items: center;
+}
+
+.panelButtons i {
+    margin-right: 5px; /* Adjust the margin as needed */
+}
+
+.interation-text{
+    font-size: 1em;
+    font-weight: 600;
+    color: #333;
+    margin-right: 5px;
+    font-family: 'Avenir', sans-serif;
+}
+
+.interation-text-toggle{
+    font-size: 1em;
+    font-weight: 600;
+    color: #333;
+    margin-right: 5px;
+    font-family: 'Avenir', sans-serif;
+}
+
+.panelButtons:hover {
+    /* background-color: #f1f1f1; */
+    color: var(--secondary-color);
+}
+
+.panelButtons:hover .interation-text{
+    color: var(--secondary-color);
+}
+
+:root{
+    --secondary-color: #2196F3;
+}
+
+
+
+
