--- thumbnail.css
@@ -1,36 +1,0 @@
-
-#container {
-    /* max-width: 50%; */
-    margin: 0 auto;
-    /* Set the transform origin to the top left corner */
-    transform-origin: top left; 
-    transform: scale(1);
-    transition: transform 0.3s ease;
-}
-
-#thumbnail-container {
-    position: fixed;
-    bottom: 10px;
-    right: 10px;
-    width: 100px;
-    height: 75px;
-    overflow: hidden;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.2);
-}
-
-#thumbnailCanvas {
-    width: 100%;
-    height: 100%;
-    display: block;
-}
-
-#thumbnailOverlay {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    pointer-events: none; /* Allow clicks to pass through the overlay to the main canvas */
-}
